Abstracted allocation in List and associatives

+++ include/etl/base/VectorTemplate.h
@@ -95,7 +95,7 @@ class Vector : public TypedVectorBase<T> {
         strategy.reserve(*this, length);
     }
 
-    uint32_t max_size() const {
+    size_t max_size() const {
         return strategy.getMaxCapacity();
     }
